убрал логику игры в отдельный класс
